requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.3
python-dotenv==1.0.1
